decouple RectorApplication from console
