+++ src/Error/ExceptionCorrector.php
@@ -5,6 +5,7 @@ namespace Rector\Error;
 use PHPStan\AnalysedCodeException;
 use Rector\Contract\Rector\RectorInterface;
 use Throwable;
+use function Safe\sprintf;
 
 final class ExceptionCorrector
 {
